♻️ Refactoring: Repeate를 room 폴더로 이동
