+++ src/main/java/gaji/service/domain/room/entity/Event.java
@@ -1,7 +1,6 @@
 package gaji.service.domain.room.entity;
 
 import gaji.service.domain.User;
-import gaji.service.domain.myRepeat.MyRepeat;
 import jakarta.persistence.*;
 import lombok.*;
